scripts/requirements-gcal.txt
google-api-python-client>=2.193.0
google-auth>=2.48.0
